pyproject.toml
[project]
name = "image-converter"
version = "0.1.0"
description = ""
authors = [
    {name = "vedaka1",email = "vladik.kuleshov.2003@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "pyinstaller (>=6.13.0,<7.0.0)",
    "pyqt6 (>=6.9.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
